Split CH9329 tests into their own file (towards #13)
Move TestCH9329Comm from tests/utils/test_communication.py to a new tests/utils/test_ch9329.py, mirroring the production-side split. The remaining test_communication.py keeps the (currently all-skipped) list_serial_ports tests under TestListSerialPorts; they're left as a TODO marker rather than deleted, since list_serial_ports itself remains in communication.py and warrants coverage once mocked correctly. That's a separate issue to raise.
